Use a custom scanner for text blocks
It doesn't seem possible to use regex to match all text in the general
content blocks up to but not including {%, {{ and {[. We have to use a
custom scanner instead that can inspect the characters one by one and
return the tokens as appropriate.

The general set up is based on the tree-sitter-python code though our
scanner doesn't seem to need any state at the moment.